requirements.txt
python-dotenv~=0.21.0
python-telegram-bot==13.15
redis==3.2.1
requests==2.31.*
retry==0.9.*
environs==9.5.*